+++ packages/array-union/index.mjs
@@ -1,7 +1,7 @@
 var arrayUnion = union;
 
 /*
-  union([1, 2, 5, 6], [2, 3, 4, 6]); // [1, 2, 3, 4, 5, 6]
+  union([1, 2, 5, 6], [2, 3, 4, 6]); // [1, 2, 5, 6, 3, 4]
 */
 
 function union(arr1, arr2) {