+++ packages/function-partial/CHANGELOG.md
@@ -1,5 +1,29 @@
 # just-partial-it
 
+## 3.4.0
+
+### Minor Changes
+
+- Rename node module .js -> .cjs
+
+## 3.3.0
+
+### Minor Changes
+
+- Consistent export for index.ts and index.d.ts
+
+## 3.2.1
+
+### Patch Changes
+
+- fix: reorder exports to set default last #488
+
+## 3.2.0
+
+### Minor Changes
+
+- package.json updates to fix #467 and #483
+
 ## 3.1.1
 
 ### Patch Changes