+++ packages/number-random-integer/CHANGELOG.md
@@ -1,5 +1,23 @@
 # just-random-integer
 
+## 4.2.0
+
+### Minor Changes
+
+- Rename node module .js -> .cjs
+
+## 4.1.1
+
+### Patch Changes
+
+- fix: reorder exports to set default last #488
+
+## 4.1.0
+
+### Minor Changes
+
+- package.json updates to fix #467 and #483
+
 ## 4.0.0
 
 - Rewritten logic. Results are now evenly distributed between entire range (inclusively). Also includes fixes for negative ranges and allows arguments to be added in either order.