+++ packages/object-is-empty/CHANGELOG.md
@@ -1,6 +1,37 @@
 # just-is-empty
 
+## 3.4.1
+
+### Patch Changes
+
+- Fix is-empty return type
+
+## 3.4.0
+
+### Minor Changes
+
+- Check objects for symbols too
+
+## 3.3.0
+
+### Minor Changes
+
+- Rename node module .js -> .cjs
+
+## 3.2.1
+
+### Patch Changes
+
+- fix: reorder exports to set default last #488
+
+## 3.2.0
+
+### Minor Changes
+
+- package.json updates to fix #467 and #483
+
 ## 3.1.2
+
 ### Patch Changes
 
 - Types