+++ packages/object-reduce/CHANGELOG.md
@@ -1,5 +1,29 @@
 # just-reduce-object
 
+## 2.3.0
+
+### Minor Changes
+
+- Rename node module .js -> .cjs
+
+## 2.2.2
+
+### Patch Changes
+
+- use default export in typings
+
+## 2.2.1
+
+### Patch Changes
+
+- fix: reorder exports to set default last #488
+
+## 2.2.0
+
+### Minor Changes
+
+- package.json updates to fix #467 and #483
+
 ## 2.1.1
 
 ### Patch Changes