+++ packages/string-kebab-case/CHANGELOG.md
@@ -1,5 +1,23 @@
 # just-kebab-case
 
+## 4.2.0
+
+### Minor Changes
+
+- Rename node module .js -> .cjs
+
+## 4.1.1
+
+### Patch Changes
+
+- fix: reorder exports to set default last #488
+
+## 4.1.0
+
+### Minor Changes
+
+- package.json updates to fix #467 and #483
+
 ## 4.0.3
 
 ### Patch Changes